requirements.txt
python-telegram-bot==20.7
openai==1.12.0
crewai==0.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.2.0
matplotlib==3.8.2 